+++ src/_BSplineManifold.jl
@@ -13,7 +13,7 @@ Construct Rational B-spline manifold from given control points and B-spline spac
 julia> using StaticArrays
 
 julia> P = BSplineSpace{2}(KnotVector(0,0,0,1,1,1))
-BSplineSpace{2, Float64}(KnotVector([0.0, 0.0, 0.0, 1.0, 1.0, 1.0]))
+BSplineSpace{2, Int64}(KnotVector([0, 0, 0, 1, 1, 1]))
 
 julia> a = [SVector(1,0), SVector(1,1), SVector(0,1)]
 3-element Vector{SVector{2, Int64}}:
